♻️ coment 컴포넌트 재사용 가능하게 분리
- 데이터 호출은 부모 컴포넌트에서 실행
